Create generic plug-in structure
Here is a generic plug-in structure that'd work for any type of program
really. It loads plug-ins dynamically from specified folders (by default
one for each plug-in type but that could be different at some point).
There it parses some metadata about the plug-in and determines if it
should be added to the application.