Allow plug-ins to require ranges of API versions
The plug-in metadata can now require that the plug-in API of specific
types of plug-ins falls within a certain range. Previously it could only
indicate the maximum API version of its own type of plug-in. Now it can
specify the version numbers of other APIs as well, and it can specify a
range. Both the maximum and minimum are required for now. Incremented
all API versions accordingly.