Make listeners into weak references
If the instance of the listener function is removed, the reference in
the list of listeners is also removed. To do this, we need to introduce
an extra layer in the list of listeners for the instance, and make it so
that if the instance is deleted, the listeners below it are also
deleted.