Rewrite model decorator
This solution is a lot shorter. It looks a lot like the original idea,
but this time supports making models out of built-in types by allowing
listening to all changes including via __setitem__ and append. It's
faster and more elegant than wrapping all attributes, and doesn't allow
for problematic cases like mutual referencing.