Space after commas
I don't usually do this, as I see the comma more as a separator than as
part of a sentence and don't see why we should make an exception for it.
But I'm trying to make the code closer to the standards for their
respective languages, and this seems very pervasive.