Remove duplicate modules
This was caused by Windows and PyCharm messing up when renaming these files to their lowercase variants. Since Windows' file system is case-preserving but not case sensitive, it just committed the lowercase and didn't remove the uppercase variant. The uppercase variant could not be deleted on Windows, since it had only the lowercase! So I had to use a different machine to even be able to delete them...