Move logger fallback code to class
This class can then actually implement the plug-in interface, though it
is not loaded by the plug-in manager. This one still needs special care
by the logger to only call the fallback when no other loggers are
available, not as with the default behaviour, to call all loggers
including this one.