Allow protected access inside wrappers
These wrapper functions are accessing attributes of self, which is
officially only given to it after initialisation, not at the time of the
function definition. This triggers PyLint to say that it's accessing
attributes of a foreign instance, which is not entirely correct here.
All these attributes are defined in the same module, yea, even the same
function.