Move is-path-check to add_plugin_location
We can give an exception earlier because it can be trusted that
add_plugin_location puts only directories in __plugin_locations. This is
only not secure if the directory is deleted in the meanwhile, but we
can't really be secure against code being deleted while the program is
running anyway.