Register completed plug-ins with their registrars
There's a problem with this: Registering a plug-in may never fail, but it needs to be in a plug-in. Since plug-ins can't be loaded according to their dependency graph (the graph may be circular), it is not possible to guarantee that all dependencies are registered before registering the dependee. Therefore it may occur that a dependee is already registered when its dependency later fails to register. In the future, we may need to unregister all dependees (and prevent any new dependees from loading) when registering fails.