Add loggertype stub
This plug-in defines the concept of logging. It is a bit of a special case for Luna since this plug-in is a very hard requirement for Luna and the entire application might crash if there is a bug in this plug-in type or its API. It is the only plug-in that will be 'trusted' code.