Move all plug-ins to plugins directory
This is to be able to define new types of plug-ins without having to modify the Luna base code to indicate where it must now search. Alternative would have been to search in the entire base directory but that would give conflicts with the luna package and might give unexpected results in the future if there are other directories with an __init__.py file, such as in the build system.