Make APIs completely static
There can be only one API for every type anyway, so all methods can be made static to prevent having to make instances of the APIs, which can introduce exceptions in a place where we don't like them. And when every method is static we might as well just use module-level functions.