Removed interface from loggertype
The metadata is already a sort of interface. But the plug-ins of a type needed to have access to the interface during initialisation in order to define a class that implements it, but at that point the plug-in type plug-in may not have been loaded yet. That is why the interfaces need to be removed and everything in the interface needs to be put in the metadata instead. The metadata validator becomes a lot more busy. But it is more beautiful this way, I think. The only real problem is the documentation of the plug-in. I'll make a separate document for the functional documentation on how to implement each plug-in type.