Remove check for existence of logger in set_levels
The set_levels function is also used by the registrar to set the default levels for a plug-in, which enters the initial key-value pair into the __logger_levels dictionary. So this is also a way to put the first entry into that dictionary, and it can't be breaking when putting something in that dictionary.