Build Python from source if not found
This is of course optional. There is a new option DEPENDENCIES_BUILD_MISSING that causes dependencies to build if they are missing. This option is disabled by default. Currently it does not work on my machine since I don't have Visual Studio 2015. Need to test it further, perhaps build it in Cygwin.