Implement can_read and can_write
These say they can read or write if the scheme is file. They use Python's file I/O, which can read from file schemes even if the file is not on the actual local hard disk but, say, a network drive or USB stick or something. So this is fine.