Add simple initial test case
This is mostly just to try out how we're going to implement unit tests within the plug-ins. This test works as long as the test is ran from within the plug-in directory with the application's main directory added to PYTHONPATH. I'm going to have to add such a construct to the CMake script.