Switch to CC0 license
The Unlicense was very nice philosophically, but in practice it may be unapplicable in some jurisdictions because it is too free, causing the Unlicense to be invalid altogether and restricting use of the software in those jurisdictions. CC0 does this a bit better and performs basically the same in practice.