Add test case for atomicity of read
Fairly complex test! It's also not entirely correct yet. In particular, the test only fails if the initial string to overwrite the original file with is longer than the original file. Also doesn't yet clean up the file after writing. But the idea I came up with to test this atomicity is quite brilliant, if I may say so myself.