Always open without buffering in tests
This technically changes the test, but you could achieve the same result by testing with writing pieces of the buffer size instead of one byte at a time. Problem is that the actual buffer size is difficult to obtain cross-platformly, and you'd have to generate test strings of that length.