Move optional BUILD_SPHINX routine into find script
Going to do this for the rest of the dependencies as well. Then we will do away with the BUILD_MISSING_DEPENDENCIES option and allow for individual dependencies to be built regardless of whether they are installed on the system (if installed as well, the custom-built version should get used).