No longer register loggers inside loggertype plug-in
This is now done by the plug-in registry itself. This reduced boilerplate quite a lot, but I worry that it removed some of the convenience of calling the logging functions all the way inside the plugins module, which is boilerplate in itself (though less boilerplate when creating the plug-in, it's a bit more when implementing the details).