Allow plug-ins to have their own CMakeLists
They now also register their tests via this CMakeLists. This way, plug-ins can stay more local and decide on their own way to include tests. For this reason, the --Tests-- category in the main CMakeLists needs to be moved higher so that testing is already enabled by the time the plug-in CMakeLists are configured.