Obtain MIME types from plug-in metadata
This executes a little bit of Python that looks inside the metadata function. It often gives an error (for every package that is not a MIME plug-in) but these errors are hidden. Only if a successful MIME type was found will it return something. Currently it just sends a message with the found MIME types but we'll expand on this later.