Move fallback logging to separate function
We only use the 'warning' level so far, so only one function is needed.

Instead of using the (neater) try-catch block, we now check first before we leap. This is because using a try-catch block would cause the 'ImportError: Can't find logger' message to be appended in case of using the fallback logger.