Add data type plug-ins
These plug-ins are mainly meant for communicating support for certain types of data. For instance, a user interface element for user settings may know how to display a text box for strings and a slider for floating point values, but has no user interface element for editing polygons and such.