Manually list all Python files
This means that from now on, every time we add or remove a Python file in the project, we must add it to the list here. That is a disadvantage.
It is an advantage though that we will now re-run the CMake script when this happens. It also happens to make testing a lot more efficient and robust because it doesn't find all the Python files in the dependencies in the build directory any more.