Support serialising/deserialising enums nested in classes
Some logic is involved in walking the path towards the enum. It works in two parts: First it keeps trying to import with longer and longer subsets of the input string until it finds a module in sys.modules. Then it walks the rest of the path using getattr to hopefully arrive at the enum instance.