Initialise output stream with bytes instead of writing to it
If we would be writing to it, the cursor would be at the end of the stream and reading from it at the same time would result in empty data. This solution is not final since we want to be able to read and write from the stream at the same time, but it works for now since we're just writing to it all at once anyway.