Implement new configuration API in configuration API itself
I'm not entirely happy yet with how these are saved and loaded, so those two methods are not yet implemented. It might also be good to change those methods to some name that doesn't start with an underscore so that the user of the API doesn't have to call methods with an underscore. But then the reserved-names requirement has to change a bit.