Implement __setitem__
The ValidatedDictionary now inherits most of its functionality from the built-in dict, but it can still be modified on the fly by the listeners since this is a Python class now. Only __setitem__ will need to be overwritten and some additional functions need to be written to allow adding and resetting items.

The class as it stands now is completely unusable since you can't add any items.