Implement deserialising validated dictionaries
This basically creates the basic file format for all config files. However I'm not quite happy with this definition metadata field, so that is kind of subject to change. We'll have to see about two things: First, the definition metadata is double data, since it should also be included as the file name. Second, the definition metadata is a huge security risk if an attacker has control over at least one configuration file, since it can overwrite any arbitrary configuration, including those from internal modules and whatnot. If a user imports a profile he wouldn't expect his preferences to be overwritten, and an application could store data in the preferences such as a URI which can then be changed to the attacker's site, and... Well, in short, this is a security risk.